+++ vector/src/main/java/im/vector/riotx/features/home/HomeActivity.kt
@@ -146,7 +146,8 @@ class HomeActivity : VectorBaseActivity(), ToolbarConfigurable {
             promptCompleteSecurityIfNeeded()
         }
 
-        shortcutsHandler.observeRoomsAndBuildShortcuts(context = this)
+        shortcutsHandler.observeRoomsAndBuildShortcuts()
+                .disposeOnDestroy()
     }
 
     private fun promptCompleteSecurityIfNeeded() {